Finished and cleaned up backend all views/urls files

+++ backend/comments/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.permissions import IsAuthenticated, AllowAny
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
 from authentication.models import User
 from .models import Comments
@@ -9,22 +9,6 @@ from .serializers import CommentsSerializer
 from django.shortcuts import get_object_or_404
 
 
-#GET Comments by user_id
-@api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def get_by_user(request, pk):
-    if request.method=='GET':
-        try:
-            user = User.objects.get(pk=pk)
-        except:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        # try:
-        #     comments = Comments.objects.filter(user_id=user_id)
-        # except comments == None:
-        #     return Response (status=status.HTTP_204_NO_CONTENT)
-        serializer = CommentsSerializer(user, many=True, data=request.data)
-        serializer.is_valid()
-        return Response(serializer.data)
 #GET All Comments
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
@@ -33,6 +17,7 @@ def get_all_comments(request):
     comments = Comments.objects.all()
     serializer = CommentsSerializer(comments, many=True)
     return Response(serializer.data)
+
 #CREATE Comment
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
@@ -43,13 +28,17 @@ def user_post_comm(request):
             serializer.save(user=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+#GET Comments by user_id
 #DELETE Comment
-@api_view(['DELETE']) #Done
+@api_view(['GET','DELETE']) #Done
 @permission_classes([IsAuthenticated]) 
-def delete_comm(request, pk):
-    if request.method == 'DELETE':
-        comment = get_object_or_404(Comments, pk=pk)
+def user_comm(request, pk):
+    comment = get_object_or_404(Comments, pk=pk)
+    if request.method == 'GET':
         serializer = CommentsSerializer(comment)
+        return Response(serializer.data)
+    if request.method == 'DELETE':
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
